+++ app/src/main/java/com/space/chatApp/data/local/MessageEntity.kt
@@ -2,7 +2,6 @@ package com.space.chatApp.data.local
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import com.space.chatApp.domain.model.UserType
 
 @Entity(tableName = "message_table")
 data class MessageEntity(
