+++ app/src/main/java/com/space/chatApp/di/RepositoryModule.kt
@@ -1,8 +1,8 @@
 package com.space.chatApp.di
 
 import com.space.chatApp.data.local.data_store.ThemeMode
-import com.space.chatApp.data.repositoryImpl.ChatRepositoryImpl
-import com.space.chatApp.data.repositoryImpl.ThemeRepositoryImpl
+import com.space.chatApp.data.repository.ChatRepositoryImpl
+import com.space.chatApp.data.repository.ThemeRepositoryImpl
 import com.space.chatApp.domain.repository.ChatRepository
 import com.space.chatApp.domain.repository.ThemeRepository
 import org.koin.dsl.module
