+++ app/src/main/java/com/space/chatApp/di/ViewModelModule.kt
@@ -1,8 +1,7 @@
 package com.space.chatApp.di
 
 import com.space.chatApp.presentation.chat_activity.viewModel.ChatActivityViewModel
-import com.space.chatApp.presentation.chat_screen.mapper.MessageDomainToUIMapper
-import com.space.chatApp.presentation.chat_screen.viewModel.ChatViewModel
+import com.space.chatApp.presentation.chat_screen.view_model.ChatViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
