--- app/src/main/java/com/space/chatApp/domain/model/UserType.kt
@@ -1,6 +1,0 @@
-package com.space.chatApp.domain.model
-
-object UserType {
-    const val FIRST_USER_ID = "FirstUser"
-    const val SECOND_USER_ID = "SecondUser"
-}
